+++ crud/common/sharding/utils.lua
@@ -1,4 +1,5 @@
 local digest = require('digest')
+local errors = require('errors')
 local msgpack = require('msgpack')
 
 local utils = {}
@@ -7,6 +8,8 @@ utils.SPACE_NAME_FIELDNO = 1
 utils.SPACE_SHARDING_KEY_FIELDNO = 2
 utils.SPACE_SHARDING_FUNC_NAME_FIELDNO = 2
 utils.SPACE_SHARDING_FUNC_BODY_FIELDNO = 3
+
+utils.ShardingHashMismatchError = errors.new_class("ShardingHashMismatchError", {capture_stack = false})
 
 function utils.extract_sharding_func_def(tuple)
     if not tuple then