+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Fix processing storage error for tuple-merger implementation of
   select/pairs (#271).
 * Do not change input tuple object in requests.
+* Add automatic reload of DDL schema (#212).
 
 ## [0.10.0] - 01-12-21